requirements.txt
pandas==2.3.0
openpyxl==3.1.5
google-genai==1.19.0
pydantic==2.11.5
python-dotenv==0.19.2
requests==2.32.4
pillow==11.2.1
pytesseract==0.3.13
PyMuPDF==1.23.26
